ignore runt sized isoch packets

+++ src/Server.cpp
@@ -485,7 +485,8 @@ int Server::L2_quintuple_filter(void) {
 
 void Server::Isoch_processing (int rxlen) {
 #ifdef HAVE_ISOCHRONOUS
-    if (rxlen > (sizeof(UDP_datagram) +  sizeof(client_hdr_v1) + sizeof(client_hdr_udp_isoch_tests))) {
+    // Ignore runt size isoch packets
+    if (rxlen > (int) (sizeof(UDP_datagram) +  sizeof(client_hdr_v1) + sizeof(client_hdr_udp_isoch_tests))) {
 	struct client_hdr_udp_isoch_tests *testhdr = (client_hdr_udp_isoch_tests *)(mBuf + sizeof(client_hdr_v1) + sizeof(UDP_datagram));
 	struct UDP_isoch_payload* mBuf_isoch = &(testhdr->isoch);
 	reportstruct->isochStartTime.tv_sec = ntohl(mBuf_isoch->start_tv_sec);
